pyproject.toml
#
# Use this configuration file to control what RobotPy packages are installed
# in your project.
#
[project]
name = "romipykitsubsystems"
version = "0.1.0"
description = "A WPILib Romi Robot"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "wpilib>=2026.2.1",
    "robotpy-pykit==1.0.2",
    "robotpy-pykit-watch==1.0.0",
    "robotpy-halsim-gui>=2026.2.1",
    "robotpy>=2026.2.1",
    "robotpy-cli",
    "black>=25.9.8",
    "pylint>=4.0.2",
    "robotpy-pathplannerlib>=2026.1.1",
    "phoenix6>=26.1.0",
    "robotpy-navx>=2026.0.0",
    "robotpy-rev>=2026.0.1",
    "robotpy-commands-v2>=2026.2.1",
    "robotpy-halsim-ws",
    "robotpy-native-romi>=2026.2.1",
    "robotpy-romi>=2026.2.1",
    "coverage>=7.13.1",
]
[tool.robotpy]

# Version of robotpy this project depends on
# to see what versions are available: https://pypi.org/pypi/robotpy/json

# Version of robotpy this project depends on
robotpy_version = "2026.2.1"

# Which extra RobotPy components should be installed
# -> equivalent to `pip install robotpy[extra1, ...]
components = [
    # "all",
    #"apriltag",
    "commands2",
    # "cscore",
    #"navx",
    #"pathplannerlib",
    # "phoenix5",
    #"phoenix6",
    # "photonvision",
    # "playingwithfusion",
    #"rev",
    "romi",
    "sim",
    # "xrp",
]

# Other pip packages to install
requires = []
